requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
discord.py==2.4.0
Django==5.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
python-decouple==3.8
sqlparse==0.5.1
yarl==1.9.4
